requirements.txt
# Core STT
faster-whisper>=1.0.0

# Audio recording
sounddevice>=0.4.6
numpy>=1.24.0

# Keyboard simulation and hotkeys
pynput>=1.7.6
pyperclip>=1.8.2

# System tray UI
pystray>=0.19.5
Pillow>=10.0.0

# Build executable
pyinstaller>=6.0.0
